% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifier.R
\name{modifier}
\alias{modifier}
\title{Create or read a set of data modification rules}
\usage{
modifier(..., .file)
}
\arguments{
\item{...}{A comma-separated list of modification rules.}

\item{.file}{A character vector of file locations.}
}
\value{
An object of class \code{modifier}.
}
\description{
Create or read a set of data modification rules
}
\examples{
m <- modifier( if (height < mean(height)) height <- 2*height
, if ( weight > mean(weight) ) weight <- weight/2  )
modify(women,m)


}
