% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifier.R, R/modify.R
\name{modify}
\alias{modify}
\alias{modify,data.frame,modifier-method}
\title{Modify a data set}
\usage{
modify(dat, x, ...)

\S4method{modify}{data.frame,modifier}(dat, x, logger = NULL, ...)
}
\arguments{
\item{dat}{A \code{data.frame}}

\item{x}{A \code{modifier} object containing modifying rules.}

\item{...}{Extra arguments.}

\item{logger}{Optional. A \code{lumberjack}-compatible logger object.}
}
\description{
Modify a data set
}
\examples{
m <- modifier( if (height < mean(height)) height <- 2*height
, if ( weight > mean(weight) ) weight <- weight/2  )
modify(women,m)
}
