% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{modify_so}
\alias{modify_so}
\title{Shortcut to modify data}
\usage{
modify_so(dat, ...)
}
\arguments{
\item{dat}{A \code{data.frame}}

\item{...}{A comma-separated list of modifying rules.}
}
\description{
Shortcut to modify data
}
