% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_yaml}
\alias{export_yaml}
\alias{origin}
\alias{origin<-}
\alias{description}
\alias{description<-}
\alias{label}
\alias{label<-}
\alias{do_by}
\alias{mean_by}
\alias{sum_by}
\alias{max_by}
\alias{min_by}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{validate}{\code{\link[validate:export_yaml]{as_yaml}}, \code{\link[validate]{description}}, \code{\link[validate:description]{description<-}}, \code{\link[validate]{do_by}}, \code{\link[validate]{export_yaml}}, \code{\link[validate]{label}}, \code{\link[validate:label]{label<-}}, \code{\link[validate:do_by]{max_by}}, \code{\link[validate:do_by]{mean_by}}, \code{\link[validate:do_by]{min_by}}, \code{\link[validate]{origin}}, \code{\link[validate:origin]{origin<-}}, \code{\link[validate:do_by]{sum_by}}}
}}

