% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\docType{package}
\name{dcmodify}
\alias{dcmodify}
\alias{dcmodify-package}
\alias{dcmodify-package}
\title{Data Modification By Modifying Rules}
\description{
Data Modification By Modifying Rules
}
\section{Introduction}{

Data often contain errors and missing data. Experts can often correct
commonly occuring errors based on simple conditional rules. This package
facilitates the expression, management, and application of such rules on data
sets.


The general workflow in \code{dcmodify} follows the following pattern.
\itemize{
  \item Define or read a set of rules with \code{\link{modifier}}. 
  \item \code{\link{modify}} data with the modification rules.
  \item Examine the results either graphically or by summary.
}

There are several convenience functions that allow one to define modification
rules from the commandline, through a (freeform or yaml) file and to
investigate and maintain the rules themselves. Please have a look at the
introductory vignette

\code{vignette("introduction",package="dcmodify")}
}

