% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifier.R, R/modify.R
\docType{methods}
\name{modify}
\alias{modify}
\alias{modify,data.frame,modifier-method}
\title{Modify a data set}
\usage{
modify(dat, x, ...)

\S4method{modify}{data.frame,modifier}(dat, x, ...)
}
\arguments{
\item{dat}{An R object carrying data}

\item{x}{An R object carrying modififying rules}

\item{...}{Options.}
}
\description{
Modify a data set
}
\examples{
m <- modifier( if (height < mean(height)) height <- 2*height
, if ( weight > mean(weight) ) weight <- weight/2  )
modify(women,m)
}
